pyproject.toml
[project]
name = "internal-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bson>=0.5.10",
    "fastapi>=0.135.3",
    "flagembedding>=1.4.0",
    "grandalf>=0.8",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "loguru>=0.7.3",
    "minio>=7.2.20",
    "modelscope>=1.36.2",
    "numpy>=2.4.4",
    "openai-agents>=0.14.7",
    "pymilvus>=2.6.11",
    "pymilvus-model>=0.3.2",
    "pymongo>=4.17.0",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.5.0",
    "uvicorn[standard]>=0.46.0",
]
